[Warhammer]
++ Starting Ability
++ Ability_infos Table
++ Action Points Regen
** Changed Environement.TickCount -> TcpManager.GetTimeStampMs(); (int to long)
Environement.TickCount becomes negative after several days

+++ WarhammerV2/Trunk/WorldServer/World/Interfaces/BaseInterface.cs
@@ -23,7 +23,7 @@ namespace WorldServer
             this.Obj = Obj;
         }
 
-        public virtual void Update(int Tick)
+        public virtual void Update(long Tick)
         {
 
         }
