+++ WarhammerV2/Trunk/WorldServer/World/Interfaces/ItemsInterface.cs
@@ -145,7 +145,7 @@ namespace WorldServer
                 }
             }
         }
-        public override void Update(int Tick)
+        public override void Update(long Tick)
         {
 
         }
