+++ WarhammerV2/Trunk/WorldServer/World/Interfaces/StatsInterface.cs
@@ -41,7 +41,7 @@ namespace WorldServer
 
             base.Load();
         }
-        public override void Update(int Tick)
+        public override void Update(long Tick)
         {
 
         }
